+++ Assignment/src/java/Controllers/AdminController/Product/delete.java
@@ -17,7 +17,10 @@ public class delete extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-
+        if(request.getSession().getAttribute("admin") == null) {
+            request.getRequestDispatcher("/admin/auth/login.jsp").forward(request,response);
+            return;
+        }
 		Product product = Product.find(Integer.parseInt(request.getParameter("id")));
 		
 		product.delete();
