+++ Assignment/src/java/Controllers/AdminController/Staff/delete.java
@@ -17,7 +17,10 @@ public class delete extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-
+        if(request.getSession().getAttribute("admin") == null) {
+            request.getRequestDispatcher("/admin/auth/login.jsp").forward(request,response);
+            return;
+        }
 		Admin admin = Admin.find(Integer.parseInt(request.getParameter("id")));
 		
 		admin.delete();
@@ -29,10 +32,5 @@ public class delete extends HttpServlet {
         JsonObjectBuilder job = Json.createObjectBuilder()
                 .add("script", "Swal.fire({title: 'Completed', text: 'Staff Delete Successfully', icon: 'success'}).then(()=>{$('#admin-table').ajax_html();})");
         out.print(job.build().toString());
-    }
-
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }   
+    } 
 }
